Split canned locales from localization boilerplate
