+++ packages/react-components/review/react-components.api.md
@@ -398,13 +398,6 @@ export interface MicrophoneButtonStrings {
     onLabel: string;
 }
 
-// @public
-export const namedLocales: Record<string, {
-    locale: Locale;
-    englishName: string;
-    displayName: string;
-}>;
-
 // @public (undocumented)
 export type OmitNever<T> = Pick<T, AllKeys<T>[keyof AllKeys<T>]>;
