+++ packages/react-components/src/locales/index.ts
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { Locale } from '../localization/';
+import _en_US from './translated/en_US.json';
+import _de_DE from './translated/de_DE.json';
+import _ar_SA from './translated/ar_SA.json';
+import _fr_FR from './translated/fr_FR.json';
+
+/** Locale for English (US) */
+export const en_US: Locale = { strings: _en_US };
+/** Locale for German (Germany) */
+export const de_DE: Locale = { strings: _de_DE };
+/** Locale for French (France) */
+export const fr_FR: Locale = { strings: _fr_FR };
+/** Locale for Arabic (Saudi Arabia) */
+export const ar_SA: Locale = { strings: _ar_SA };
