+++ packages/react-components/src/locales/translated/ar_SA.json
