+++ packages/react-components/src/localization/index.ts
@@ -2,4 +2,3 @@
 // Licensed under the MIT license.
 
 export * from './LocalizationProvider';
-export * from './locales';
